update Getting started as an optional inner field for integration

+++ integrations/observability/amazon_vpc_flow/assets/create_table_vpc_schema-1.0.0.sql
@@ -6,7 +6,7 @@ CREATE EXTERNAL TABLE IF NOT EXISTS {table_name} (
     dstaddr string,
     srcport int,
     dstport int,
-    protocol int,
+    protocol bigint,
     packets bigint,
     bytes bigint,
     start bigint,
